Update validate() to return result with valid message
This commit adds the "message" property to ValidationResult, and changes the signature of isValid and assertValid.

This solves the following problems:

1. Asserts and guards require the asserted type to extend the input type. We cannot guarantee that with generics. The options are to assert an intersection type (which causes odd behavior for users), or to accept an anonymous `Message` (which won't catch programmer errors that should be trivial to catch at compile time). Because `validate()` is the preferred API now, we're okay with accepting anonymous `Message` for `isValid` and `assertValid` as a trade-off.

2. Assertion functions require a type annotation on validator. The method `validate` does not have this constraint, because it is not an assertion function.

The functions `isValid` and `assertValid` are methods on `Validator` now. This allows us to keep the fast-path for isValid via failFast without exposing it through an option on validate() *and* createValidator(). As a follow-up, we should decide whether this is really the right call.