Support referencing upstream branches
This project refences upstream protovalidate and protovalidate-testing in multiple places. This PR consolidates the references, and adds support for referencing a branch, in addition to a version.

This changes allows to run code generation and conformance tests from an upstream feature branch without the need for an upstream release, while minimizing the necessary temporary changes.