Release `0.6.0` (#67)
This release is compatible with the
[v0.14.0](https://github.com/bufbuild/protovalidate/releases/tag/v0.14.0)
release of Protovalidate.


## Breaking changes


We want validation rules to be easy to understand for authors and
consumers. To simplify Protovalidate, we are renaming an option and
remove two others. This is a breaking change, and you will have to
update your Protobuf files if they use the relevant options :

- `IGNORE_IF_UNPOPULATED` is renamed to `IGNORE_IF_ZERO_VALUE`.
See https://github.com/bufbuild/protovalidate/pull/397 for details.
- `(buf.validate.message).disabled` is removed.
You can replace it by adding `IGNORE_ALWAYS` to every field of the
message. See https://github.com/bufbuild/protovalidate/pull/394 for
details.
- `IGNORE_IF_DEFAULT_VALUE` is removed.
In most cases, you can replace it with `IGNORE_IF_ZERO_VALUE`. See
https://github.com/bufbuild/protovalidate/pull/396 for details.


## New Contributors
* @jrinehart-buf made their first contribution in
https://github.com/bufbuild/protovalidate-es/pull/56

**Full Changelog**:
https://github.com/bufbuild/protovalidate-es/compare/v0.5.0...v0.6.0

Signed-off-by: Sri Krishna <skrishna@buf.build>